requirements.txt
aiohttp==3.9.1
asyncio==3.4.3
python-dotenv==1.0.0
pydantic==2.5.2
dataclasses-json==0.6.3
schedule==1.2.0
openai==1.6.1
anthropic==0.7.8
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0